# doc/openapi.yml
openapi: 3.0.0
info:
  title: "API"
  version: "1.0.0-alpha"
tags:
  - name: SSTS API
    description: Identification API by facial recognition
paths:
  /:
    get:
      operationId: index
      description: API Greeting Point
      summary: Get a greeting message
      tags: ['Index' ]
      responses:
        '200':
          description: The greeting message
          content:
            text/plain:
              schema:
                type: string
                example: "The server is running ╰(*°▽°*)╯"

  /accounts:
    get:
      operationId: get_all_accounts
      description: Get all accounts
      summary: Get all accounts
      tags: [ 'Accounts' ]
      responses:
        '200':
          description: The list of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccountBase'

    post:
      operationId: create_accounts
      description: Create new accounts
      summary: Create new accounts
      tags: [ 'Accounts' ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/AccountCreate'
      responses:
        '201':
          description: The list of created accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccountBase'

    put:
      operationId: update_accounts
      description: Updates accounts
      summary: Updates accounts
      tags: [ 'Accounts' ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/Account'
      responses:
        '200':
          description: The list of updated accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccountBase'
        '404':
          description: 'Account(s) not found'
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Account(s) not found"
                  affected:
                    type: array
                    items:
                      type: string
                      format: uuid

    delete:
      operationId: delete_accounts
      description: Deletes accounts
      summary: Deletes accounts
      tags: [ 'Accounts' ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/Id'
      responses:
        '200':
          description: The list of deleted accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccountBase'
        '404':
          description: 'Account(s) not found'
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Account(s) not found"
                  affected:
                    type: array
                    items:
                      type: string
                      format: uuid

  /accounts/{id}:
    get:
      operationId: get_account
      description: Get an account by it's id
      summary: Get account by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Retrieved account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '404':
          description: 'Account not found'

  /face/describe:
    post:
      operationId: recognize_face
      description: Returns recognition information from face(s) image
      summary: Recognize face(s) from an image
      tags: [ 'Face' ]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
      responses:
        '200':
          description: The list of recognized faces encodings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FaceInfo'

  /face/compare:
    post:
      operationId: compare_faces
      description: Compares multiple faces by to a face by their encodings
      summary: Compare faces by their encodings
      tags: [ 'Face' ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FaceComparisonRequest'
      responses:
        '200':
          description: the comparison result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FaceComparisonResult'

components:
  schemas:
    FaceComparisonResult:
      type: object
      properties:
        toCheckEncoding:
          $ref: '#/components/schemas/128dArray'
        matches:
          type: array
          items:
            $ref: '#/components/schemas/EncodingsMatchInfo'

    EncodingsMatchInfo:
      type: object
      properties:
        encodings:
          $ref: '#/components/schemas/128dArray'
        isMatches:
          type: boolean
          example: true
        distance:
          type: number
          format: float

    FaceComparisonRequest:
      type: object
      properties:
        knownEncodings:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/128dArray'
        toCheckEncoding:
          $ref: '#/components/schemas/128dArray'
        tolerance:
          type: number
          format: float
          default: 0.7

    FaceInfo:
      allOf:
        - $ref: '#/components/schemas/FaceRecognitionEncodings'
      properties:
        boundingBox:
          $ref: '#/components/schemas/BoundingBox'
        imageSourceUrl:
          type: string
          format: uri
          example: "https://example.com/image.jpg"

    BoundingBox:
      allOf:
        - $ref: '#/components/schemas/LTBRRectangle'

    LTBRRectangle:
      type: object
      properties:
        left:
          type: integer
        top:
          type: integer
        bottom:
          type: integer
        right:
          type: integer

    FaceRecognitionEncodings:
      type: object
      properties:
        encodings:
            $ref: '#/components/schemas/128dArray'

    128dArray:
      type: array
      items:
        type: number
        format: float
        minItems: 128
        maxItems: 128

    Account:
      allOf:
        - $ref: '#/components/schemas/AccountBase'
        - $ref: '#/components/schemas/Password'
      properties:
        facesInfo:
          type: array
          items:
            $ref: '#/components/schemas/FaceInfo'

    AccountBase:
      allOf:
        - $ref: '#/components/schemas/Id'
        - $ref: '#/components/schemas/Username'
        - $ref: '#/components/schemas/Email'

    AccountCreate:
      allOf:
        - $ref: '#/components/schemas/Username'
        - $ref: '#/components/schemas/Email'
        - $ref: '#/components/schemas/Password'

    Id:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426655440000"

    Username:
      type: object
      properties:
        username:
          type: string
          example: "john_doe"

    Email:
      type: object
      properties:
        email:
          type: string
          format: email
          example: "john@anonymous.user"

    Password:
      type: object
      properties:
        password:
          type: string
          example: "super duper hyper mega secret password you'll never guess"